// temp/pms/models/com_ever_edu_global_s3_dto_res_S3ObjectListResDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type com_ever_edu_global_s3_dto_res_S3ObjectListResDto = {
    bucketName?: string;
    key?: string;
    lastModified?: string;
    etag?: string;
    size?: number;
};